rename file names to add cuda prefix
